flask/requirements.txt
Flask
flask-cors
flask-session
pyarango
graphql-server[flask]
graphene
python-arango
Pillow
pandas
